Boston Data Analysis
The Boston Housing Dataset consists of price of houses in various places in Boston. Alongside with price, the dataset also provide information such as Crime (CRIM), areas of non-retail business in the town (INDUS), the age of people who own the house (AGE), and many other attributes. The dataset itself is available here (https://www.kaggle.com/c/boston-housing/data). However, because we are going to use scikit-learn, we can import it right away from the scikit-learn itself.